+++ Problems/P/P_DFA/Solvers/DFASolver.cs
@@ -1,9 +1,9 @@
 ﻿using API.Interfaces;
 using API.Interfaces.Graphs;
-using API.Problems.NPComplete.NPC_DFA;
+using API.Problems.P.P_DFA;
 using Xunit;
 
-namespace API.Problems.NPComplete.NPC_DFA.Solvers;
+namespace API.Problems.P.P_DFA.Solvers;
 
 public class DFASolver : ISolver<DFA>
 {
